split read db interaction
